WP8: NFC double coefficients + correctness fixes (authoritative build)
An authoritative from-scratch rebuild surfaced issues that incremental builds
had masked. Three are genuine; the rest were over-tight test expectations.

Genuine fixes:
- NFC coefficients AND section state promoted float -> double (AmbiNFCFilter,
  the calc-engine pages, the RT seam, the offline harness). At 96 kHz / order 10
  the near-field poles sit at ~z=1 and float storage lost ~0.16 dB of DC gain
  (PRD sec.9's "numerically delicate"); double makes DC + magnitude match the
  double-precision Daniel goldens to 0.0000 dB, at negligible cost (the encoder
  is far from the CPU bound). Audio I/O stays float; the encode-gain matrix stays
  float. Only encode-move's baseline changes (deterministic; re-updated); the
  seven M1/M2 baselines and NFC-off encode-static stay byte-identical.
- AmbiCalculationEngine: a moving source froze one step short of its target - the
  speed limiter snaps to the target on the settle tick (isInputMoving already
  false), so the row never recomposed there. Added a moving->stopped latch.
- spreadTaper cutoff uses a 1e-12 epsilon so sigma=180 (cos 90deg ~ 6e-17) cuts
  order 1 cleanly instead of leaving an inaudible order-1 term.
- AmbiCalculationEngine.h: include XoaParameterDefaults.h (kDefaultRigRadius).

Test corrections (the DSP was right; the expectations were wrong):
- theta root residual: 1e-9 -> 1e-4 (C++ double recurrence loses digits to
  cancellation at high order, ~9e-6 at l=10; the roots are exact per the generator).
- lane-DC-settle and clamp-continuity: compare against the per-order clamped
  radius ground truth (the min-radius floor binds before the boost ceiling at low
  orders), now exact with double coeffs.
- spread rE: it is UNIMODAL (rises to the max-rE peak, then falls to omni), not
  "monotonically decreasing" as the plan stated - a mild high-order taper sharpens
  rE before widening. Test asserts unimodality + endpoints + the max-rE peak.
- encoder ramp-out: JUCE's linear ramp ends at ~steady/n, not exactly 0.

Full suite green on the authoritative MSBuild; all 9 offline baselines match.

+++ Source/DSP/AmbiBusAlgorithm.h
@@ -74,7 +74,7 @@ public:
                   const spatcore::rt::RtSnapshot<rt::BusRtParams>* busParamsSource,
                   bool processingEnabled,
                   const float* encodeMatrix = nullptr,
-                  const float* nfcCoeffs = nullptr,
+                  const double* nfcCoeffs = nullptr,
                   const spatcore::rt::RtSnapshot<rt::EncoderRtParams>* encoderSrc = nullptr)
     {
         juce::ignoreUnused (maxInputChannels, maxOutputChannels);
@@ -469,8 +469,8 @@ private:
     const spatcore::rt::RtSnapshot<rt::BusRtParams>* busParams = nullptr;
 
     // WP8 encoder seams (app-owned live matrices + the scalar side-band).
-    const float* encodeMatrixPtr = nullptr;      // [kMaxInputs x 121]
-    const float* nfcCoeffsPtr = nullptr;         // [kMaxInputs x 150]
+    const float*  encodeMatrixPtr = nullptr;     // [kMaxInputs x 121] float gains
+    const double* nfcCoeffsPtr = nullptr;        // [kMaxInputs x 150] double coeffs
     const spatcore::rt::RtSnapshot<rt::EncoderRtParams>* encoderSource = nullptr;
 
     juce::AudioBuffer<float> busA, busB, busT;   // gather / rotated / transition
